Move logic to queryInputFactory
